Small moves out of core-android of things which were not really shared
