+++ core-android/src/main/java/com/jraska/github/client/core/android/OnAppCreate.kt
@@ -4,9 +4,4 @@ import android.app.Application
 
 interface OnAppCreate {
   fun onCreate(app: Application)
-
-  /**
-   * Higher priority will be executed first
-   */
-  fun priority(): Int = 0
 }
