--- core-android/src/main/java/com/jraska/github/client/core/android/View.kt
@@ -1,11 +1,0 @@
-package com.jraska.github.client.core.android
-
-import android.view.View
-
-fun View.visible(visible: Boolean) {
-  visibility = if (visible) {
-    View.VISIBLE
-  } else {
-    View.GONE
-  }
-}
