+++ feature/users/src/main/java/com/jraska/github/client/users/UserDetailViewModel.kt
@@ -7,7 +7,7 @@ import com.jraska.github.client.Owner
 import com.jraska.github.client.analytics.AnalyticsEvent
 import com.jraska.github.client.analytics.EventAnalytics
 import com.jraska.github.client.common.lazyMap
-import com.jraska.github.client.core.android.rx.toLiveData
+import com.jraska.github.client.users.rx.toLiveData
 import com.jraska.github.client.navigation.Navigator
 import com.jraska.github.client.navigation.Urls
 import com.jraska.github.client.rx.AppSchedulers
